requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
psycopg2-binary
django-oauth-toolkit
djangorestframework
django-rest-swagger
django-filter
rest_condition
Pillow
boto3
django-storages
django-multiselectfield
gunicorn
django-cors-headers
pyzbar
pyqrcode
pypng
django-resized
django-cities-light
drf-yasg
packaging
django-push-notifications